+++ src/powertree/__init__.py
@@ -1,6 +1,6 @@
 ﻿"""PowerTree - electronic circuit power tree analysis tool."""
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 APP_NAME = "PowerTree"
 FILE_EXT = ".ptproj"
 FILE_FORMAT_VERSION = 1